Screen loading and change in all css files
Add a new screen loading while the user data is loading.
Change little changes all of the styles of the project to improve the
design app.
Delete unnecessary svg files.
Change the name of the svg files.
Extract the svg of the html files into a svg files.
Delete unnecessary images.